Add suppport for protected branches + fix imports + refactor a bit
